#134 file refactoring with new include logic
search path
